tests: AppChain, account-abstraction, privacy-pools, blockchain-processing
Integration coverage across the AppChain follower / sequencer / anchoring, the
account-abstraction AppChain, the privacy-pools circuits and blockchain-processing mapping,
aligned with the chain/sync changes -- including explicit fixture disposal so the embedded
dev-chain database does not leak temp directories between runs.

+++ tests/Nethereum.AccountAbstraction.AppChain.IntegrationTests/E2E/Sequencer/SequencerAAIntegrationTests.cs
@@ -454,7 +454,7 @@ namespace Nethereum.AccountAbstraction.AppChain.IntegrationTests.E2E.Sequencer
             var latestBlock = await _fixture.Sequencer.GetLatestBlockAsync();
 
             Assert.NotNull(latestBlock);
-            Assert.Equal(blockNumber, latestBlock.BlockNumber);
+            Assert.Equal(blockNumber, (BigInteger)latestBlock.BlockNumber);
             Assert.NotNull(latestBlock.StateRoot);
             Assert.NotNull(latestBlock.TransactionsHash);
             Assert.Equal(SequencerAAFixture.CHAIN_ID, _fixture.AppChain.Config.ChainId);
